fix: metamask flask 10.17. breaking change fix
